% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tuning.R
\name{calc_pve}
\alias{calc_pve}
\title{Calculate percentage of variance explained (PVE)}
\usage{
calc_pve(object)
}
\arguments{
\item{object}{Object of class inheriting from \code{"sgspls"}. 
The function will retrieve some key parameters stored in that object.}
}
\value{
\code{PVE} returns a vector of percentage of variance explained by the scores in the PLS method.
See references for details.
}
\description{
Function to evaluate the PEV and recommend a number of components to select.
}
\examples{

set.seed(1)
n = 50; p = 500; 

X = matrix(rnorm(n * p), ncol = p, nrow = n)

beta <- rep(0,p)
beta[1:20] <- 1:20

y = X \%*\% beta + 0.1*rnorm(n)

model <- sgspls(X, y, ncomp = 3, mode = "regression")
calc_pve(model)

}
\references{
Liquet Benoit, Lafaye de Micheaux, Boris Hejblum, Rodolphe
  Thiebaut. A group and Sparse Group Partial Least Square approach applied in
  Genomics context. \emph{Submitted}.
}
\seealso{
Tuning functions \code{\link[sgspls]{calc_pve}},
\code{\link[sgspls]{tune_sgspls}}, \code{\link[sgspls]{tune_groups}}. 
Model performance and estimation  \code{\link[sgspls]{predict}}, \code{\link[sgspls]{perf.sgspls}}, \code{\link[sgspls]{coeff.sgspls}}
}
