% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{e.norm}
\alias{e.norm}
\title{Compute the euclidian norm}
\usage{
e.norm(x, exact = FALSE)
}
\arguments{
\item{x}{A vector of numeric variables}
}
\value{
Output will be a length-one numeric.
}
\description{
Computes the L2-norm of a vector \code{x}. A float 1e-16 is returned if the
exact value is 0 so that division in the PLS algorithm is well defined.
}
