% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internalPLS.R
\name{get_alphas}
\alias{get_alphas}
\title{Get the alpha values corresponding to the required sparsity}
\usage{
get_alphas(indiv_sparsity, subgroup_sparsity)
}
\arguments{
\item{indiv_sparsity}{matrix of values between 0 and 1 for the individual sparsity penalty per component}

\item{subgroup_sparsity}{matrix of values between 0 and 1 for the subgroup sparsity penalty per component}
}
\description{
Checks the sparsity levels are correct and returns 
corresponding alpha values. Alpha values should add to one.
The values correspond to the percentage of penalty put on each 
of the penalised terms in the sgspls penalty.
Current implementation forces some penalisation on the group component.
}
\examples{

# Return the alphas used for the required sparsities
get_alphas(indiv_sparsity = c(0.4,0), subgroup_sparsity = c(0.2,1))

}
