% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{nonZero}
\alias{nonZero}
\title{Return the nonzero columns and rows}
\usage{
nonZero(Z, thresh = 1e-06, zero.col = NULL)
}
\arguments{
\item{Z}{A matrix of variables, where some columns or rows are entierly zero.}

\item{thresh}{A threshold value to round values off at.}

\item{zero.col}{A vector of columns to take for the removal of the matrix.}
}
\description{
Quick function for finding the nonzero columns of a matrix.
}
\examples{
a <- matrix(c(rep(0,3),1,0,1,rep(0,3)),3,3)
a
nonZero(a) #removes all rows and columns with nonzero elements
nonZero(a,zero.col = 2) # removes only nonzero rows and columns from row 2.
}
