% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perf.R
\name{perf.sgspls}
\alias{perf.sgspls}
\title{Performance evaluation of sgsPLS objects}
\usage{
\method{perf}{sgspls}(object, validation = c("Mfold", "loo"), folds = 10,
  BIC = FALSE, progressBar = TRUE, setseed = FALSE, scale_resp = FALSE)
}
\arguments{
\item{object}{Object of class inheriting from \code{"sgspls"}. 
The function will retrieve some key parameters stored in that object.}

\item{validation}{What kind of cross validation to use, matching one of \code{"Mfold"} or 
\code{"loo"} (leave one out). Default is \code{"Mfold"}.}

\item{folds}{Number of folds to use in the cross validation.}

\item{BIC}{Return the BIC type criterion for large sample sizes (see paper for details). 
Note that this is not an actual BIC criterion.}

\item{progressBar}{Logical to show a progress bar while computing the performances}

\item{scale_resp}{Logical to scale the responses. This is useful if comparing the fit on 
multiple responses.}
}
\value{
\code{perf} returns a list that contains the following performance measures: 

\item{MSEP}{A matrix of Mean Square Error of Prediction (MSEP) estimates by cross validation. The penalty is defined
as: \deqn{MSEP = 1/n \sum \sum (f_k(x_i) - y_i)^2} see the references for details. }
\item{PRESS0}{A vector of cross validated Predictive Residual Sum of Squares (PRESS) values. 
Each column corresponds to a response. Matches the PLS package.}
\item{PRESS}{A matrix of cross validated Predictive Residual Sum of Squares (PRESS) values. 
Each row contains the values for a different component and each column corresponds to a response.} 
\item{R2}{a matrix of \eqn{R^2} values of the \eqn{Y}-variables with \code{ncomp} components} 
\item{BIC}{A BIC type criterion for large samples (see references for details). 
Note that this is not an actual BIC criterion.} 
\item{cvPred}{an array with the cross-validated predictions.}
\item{folds}{A list of the folds used in the cross validation.}
}
\description{
Function to evaluate the performance of the fitted the PLS models using various criteria. 
Evaluation is made for each component in the PLS object and can only be evaluated for regression PLS.
}
\examples{

set.seed(1)
n = 50; p = 500; 

size.groups = 30; size.subgroups = 5
groupX <- ceiling(1:p / size.groups)
subgroupX <- ceiling(1:p / size.subgroups)

X = matrix(rnorm(n * p), ncol = p, nrow = n)

beta <- rep(0,p)
betaG <- c(bSG, b0, bSG, b0, bSG, b0)
beta[1:size.groups] <- betaG

y = X \%*\% beta + 0.1*rnorm(n)

model <- sgspls(X, y, ncomp = 3, mode = "regression", keepX = 1,
                groupX = groupX, subgroupX = subgroupX,
                indiv_sparsity_x = 0.8, subgroup_sparsity_x = 0.15)

model_perf <- perf(model, folds = 5)

# Check model performance
model_perf$MSEP

}
\references{
Mevik, Bjørn-Helge, and Henrik René Cederkvist. 2004. 
  Mean Squared Error of Prediction (MSEP) Estimates for Principal Component Regression (PCR) 
  and Partial Least Squares Regression (PLSR). 
  \emph{Journal of Chemometrics} \bold{18} (9). John Wiley & Sons, Ltd.:422–29.
}
\seealso{
Tuning functions \code{\link{calc_pve}},
\code{\link{tune_sgspls}}, \code{\link{tune_groups}}. 
Model performance and estimation  \code{\link{predict}}, \code{\link{perf.sgspls}}, \code{\link{coeff.sgspls}}
}
