% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{plot.cv.sgspls}
\alias{plot.cv.sgspls}
\title{Plot the cross validation curves of a sgspls.cv object}
\usage{
\method{plot}{cv.sgspls}(obj, verbose = T)
}
\arguments{
\item{obj}{fitted sgspls.cv object}
}
\description{
Produces a plot of the cross validation curves for a "sgspls.cv" object.
}
\seealso{
sgspls.tune, perf
}
