% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict.sgspls}
\alias{predict.sgspls}
\title{Predict Method for sgspls}
\usage{
\method{predict}{sgspls}(object, newdata, ...)
}
\arguments{
\item{object}{Object of class inheriting from \code{"sgspls"}.}

\item{newdata}{Data matrix in which to look for for explanatory variables to be used for prediction.}

\item{...}{Not currently used.}
}
\value{
\code{perf} returns a list that contains the following performance measures: 
\code{predict} function produces predicted values, obtained by evaluating the sparse group subgroup PLS. 
The prediction values are calculated based on the regression coefficients of \code{object$Y} onto \code{object$variates$X}.
}
\description{
Predicted values based on sparse group subgroup PLS. New responses are predicted using a fitted model and a new matrix of observations.
}
