% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{scale_vec}
\alias{scale_vec}
\title{Compute scaled vector multiplicaiton}
\usage{
scale_vec(u, scale = 1)
}
\arguments{
\item{u}{p vector of numeric variables}

\item{scale}{a double to scale the vector by.}
}
\value{
Output will be a numeric matrix or vector.
}
\description{
Return a vector scaled by a given double
return 0 vector if scale is zero (avoid division by zero).
}
