\name{sgspls-package}
\alias{sgspls-package}
\docType{package}
\title{
\packageTitle{sgspls}
}
\description{
\packageDescription{sgspls}
}
\details{

The sgPLS package provides sparse, group and sparse group version for PLS approaches. The main functions are: sPLS for sparse PLS, gPLS for group PLS and sgPLS for sparse group PLS.
}
\author{
\packageAuthor{sgspls}

Maintainer: \packageMaintainer{sgspls}
}
\references{
Liquet Benoit, Lafaye de Micheaux, Boris Hejblum, Rodolphe
  Thiebaut. A group and Sparse Group Partial Least Square approach applied in
  Genomics context. \emph{Submitted}.

  L\^e Cao, K.-A., Martin, P.G.P., Robert-Grani\'e, C. and Besse, P. (2009).
  Sparse canonical methods for biological data integration: application to a
  cross-platform study. \emph{BMC Bioinformatics} \bold{10}:34.

  Shen, H. and Huang, J. Z. (2008). Sparse principal component analysis via
  regularized  low rank matrix approximation. \emph{Journal of Multivariate
  Analysis} \bold{99}, 1015-1034.

  Tenenhaus, M. (1998). \emph{La r\'egression PLS: th\'eorie et pratique}.
  Paris: Editions Technic.

}
\keyword{ package }
\seealso{
Tuning functions \code{\link{sgsPLS.tune.groups}},
\code{\link{sgsPLS.tune}}, \code{\link{predict}}, \code{\link{perf}} and
functions from \code{mixOmics} package: \code{summary}, \code{plotIndiv},
\code{plotVar}, \code{plot3dIndiv}, \code{plot3dVar}.
}
