% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tuning.R
\name{tune_groups}
\alias{tune_groups}
\title{Compute cross-validated mean squared prediction error for sgspls regression}
\usage{
tune_groups(parameters, group_seq = NULL, block = "X", folds = 10,
  scale_resp = TRUE, setseed = 1)
}
\arguments{
\item{parameters}{List of parameters to use in the current PLS object (see examples below).}

\item{group_seq}{a vector containing the number of groups to tune over.}

\item{block}{A string either "X" or "Y" to indicate which block to tune parameters over.}

\item{folds}{The number of folds to use in cross validation.}

\item{scale_resp}{Logical, the MSEP is standardised across responses (see perf function for details).}

\item{setseed}{False, or integer for replicating tuning parameters.}
}
\value{
\code{tune_groups} returns a list with class \code{cv.sgspls} with measures:
\item{result_tuning}{A matrix containing the tuning parameters and MSEP values.}
\item{best}{A vector containing the optimal tuning parameters. }
\item{parameters}{A list of the parameters for a sgspls object.}
\item{tuning_sparsities}{A matrix of group, subgroup and individual sparsities tuned over.}
\item{folds}{Number of folds used in cross validation.}
\item{min_cv}{Minimum MSEP score.}
\item{group_seq}{Groups tuned over in cross validation.}
}
\description{
Tuning function for finding the number of groups to select for an sgspls object.
Function is consistent in the sparsity parameters but may be tuned over the number of 
groups.
}
\examples{

 set.seed(1)
 n = 50; p = 510; 
 size.groups = 30; size.subgroups = 5
 groupX <- ceiling(1:p / size.groups)
 subgroupX <- ceiling(1:p / size.subgroups)
 
 X = matrix(rnorm(n * p), ncol = p, nrow = n)
 
 beta <- rep(0,p)
 bSG <- -2:2; b0 <- rep(0,length(bSG))
 betaG <- c(bSG, b0, bSG, b0, bSG, b0)
 beta[1:size.groups] <- betaG
 
 y = X \%*\% beta + rnorm(n)
 
 #--------------------------------------#
 #-- Set up a basic model to tune --#
 
 cv_pls <- list(X=X, Y=y, groupX=groupX, subgroupX=subgroupX, ncomp = 1:3)
 
 cv_pls_comp1 <- tune_groups(parameters = cv_pls, group_seq = 1:2, scale_resp = FALSE)
 
 #-- MSEP is on the original scale for the response --#
 cv_pls_comp1$MSEP
 cv_pls_comp1$keep_opt
 cv_pls_comp1$ncomp_opt
 
}
\references{
Liquet Benoit, Lafaye de Micheaux, Boris Hejblum, Rodolphe
  Thiebaut. A group and Sparse Group Partial Least Square approach applied in
  Genomics context. \emph{Submitted}.
}
\seealso{
\code{\link[sgspls]{sgspls}} Tuning functions \code{\link[sgspls]{calc_pve}}, \code{\link[sgspls]{tune_sgspls}}. 
Model performance and estimation  \code{\link[sgspls]{predict.sgspls}}, \code{\link[sgspls]{perf.sgspls}}, \code{\link[sgspls]{coef.sgspls}}
}
